Move methods from Arguments to more appropriate modules
Also simplifies the code somewhat.
