+++ lib/ripper_ruby_parser/sexp_handlers/method_calls.rb
@@ -4,6 +4,24 @@ module RipperRubyParser
   module SexpHandlers
     # Sexp handlers for method calls
     module MethodCalls
+      def process_args_add_star(exp)
+        generic_add_star exp
+      end
+
+      def process_args_add_block(exp)
+        _, regular, block = exp.shift 3
+        args = process(regular)
+        args << s(:block_pass, process(block)) if block
+        s(:arglist, *args.sexp_body)
+      end
+
+      def process_arg_paren(exp)
+        _, args = exp.shift 2
+        return s() if args.nil?
+
+        process(args)
+      end
+
       def process_method_add_arg(exp)
         _, call, parens = exp.shift 3
         call = process(call)
